refactor: make builder for clients
